requirements.txt
aiogram
asyncio==3.*
python-dotenv
requests